+++ semcore/date-picker/src/components/InputTrigger.jsx
@@ -524,6 +524,7 @@ const MaskedInput = ({
     placeholders.month.length !== 1 ||
     placeholders.day.length !== 1
   ) {
+    // eslint-disable-next-line no-console
     console.error({ placeholders });
     throw new Error(
       `InputTrigger placeholder prop should contain fields year, month and day each one with string value of single character length. [see above what was received] (${placeholders})`,