+++ stories/components/data-table/advanced/data-table.advanced.stories.tsx
@@ -3,10 +3,13 @@ import type { Meta, StoryObj } from '@storybook/react-vite';
 
 import AccordionInMergedRowsExample, { accordionMergedProps } from './examples/accordion_in_merged_rows';
 import AccordionWithStickyRowsExample, { accordionStickyProps } from './examples/accordion_with_sticky_rows';
+import BigTableWithStickyHeaderExample from './examples/big_table_with_sticky_header';
 import FakeMultiLineHeaderExample from './examples/fake-multi-level-header';
+import FixedColumnsWidthWithShadowsExample from './examples/fixed_columns_width_with_shadows';
 import FixedColumnsWithDiffWidthsExample from './examples/fixed_columns_with_diff_widths';
 import LinkInTableExample from './examples/link_in_table';
 import OverflowInCellsExample from './examples/overflow_in_cells';
+import OverlapCellsExample from './examples/overlap_cells';
 import RenderCellRawValueExample from './examples/render_cell_raw_data';
 import RowCellStatesExample from './examples/row_cell_states';
 import SideIndentsExample from './examples/side-indents';
@@ -35,6 +38,10 @@ export const AccordionWithStickyRows: StoryObj<typeof accordionStickyProps> = {
   args: accordionStickyProps,
 };
 
+export const BigTableWithStickyHeader: Story = {
+  render: BigTableWithStickyHeaderExample,
+};
+
 export const FixedColumnsWithDiffWidths: Story = {
   render: FixedColumnsWithDiffWidthsExample,
 };
@@ -55,10 +62,18 @@ export const OverflowInCells: Story = {
   render: OverflowInCellsExample,
 };
 
+export const OverlapCells: Story = {
+  render: OverlapCellsExample,
+};
+
 export const FakeMultiLineHeader: Story = {
   render: FakeMultiLineHeaderExample,
 };
 
 export const RenderCellRawValue: Story = {
   render: RenderCellRawValueExample,
+};
+
+export const FixedColumnsWidthWithShadows: Story = {
+  render: FixedColumnsWidthWithShadowsExample,
 };