+++ stories/components/data-table/tests/examples/accordion-tests/table-in-table-with-sorting.tsx
@@ -65,7 +65,7 @@ const ChartExample = () => {
         { name: 'cpc', children: 'CPC' },
         { name: 'vol', children: 'Vol.', gtcWidth: '100px' },
       ]}
-      expandedRows={new Set()}
+      expandedRows={new Set<string>()}
       renderCell={(props) => {
         if (props.columnName === 'vol') {
           return (