+++ tools/changelog-handler/src/parsers/releaseChangelogParser.ts
@@ -79,6 +79,8 @@ export const releaseChangelogParser = (
         );
       }
       traversingComponent = component;
+    } else if (token.type === 'heading' && token.level === 4) {
+      // just skip changelogs for major
     } else if (
       traversingVersion &&
       traversingDate &&
@@ -92,10 +94,14 @@ export const releaseChangelogParser = (
           );
         }
 
+        if (changelogFilePath.includes('semcore/ui/CHANGELOG.md')) {
+          continue;
+        }
+
         const prefix = item.text[0];
         const restText = item.text.slice(1);
 
-        if (typeof prefix === 'string' || prefix.type !== 'strong') {
+        if ((typeof prefix === 'string' || prefix.type !== 'strong')) {
           throw new Error(
             `Invalid prefix for changelog change. Expected strong text, got ${JSON.stringify(
               prefix,