+++ semcore/base-components/CHANGELOG.md
@@ -2,12 +2,21 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.2.0] - 2025-08-08
+## [16.2.1] - 2025-08-29
+
+### Changed
+
+- Type description for `PopperPopperProps`/`PopperProps`/`ScrollAreaProps`/`ScrollBarProps`/`BoxProps`.
 
 ### Added
 
 - `useEllipsis` hook.
 - `Hint` component.
+
+## [16.2.0] - 2025-08-08
+
+### Added
+
 - `text-align` CSS property to `Box` component.
 
 ## [16.1.2] - 2025-07-23