+++ semcore/breadcrumbs/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.4.1] - 2025-08-29
+
+### Changed
+
+- Type description for `BreadcrumbsProps` & `BreadcrumbsItemProps`.
+
+## [16.4.0] - 2025-08-07
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/icon` [16.4.0 ~> 16.5.0], `@semcore/base-components` [16.1.2 ~> 16.2.0]).
+
 ## [16.3.0] - 2025-07-23
 
 ### Changed