+++ semcore/bulk-textarea/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.4.0] - 2025-08-29
+
+### Added
+
+- Controlled mode for errors.
+- `onImmediatelyChange` internal property.
+
+### Fixed
+
+- Incorrect paddings for line numbers starting from 1000 lines.
+
+## [16.3.1] - 2025-08-29
+
+### Changed
+
+- Type description for `BulkTextareaProps`.
+
+## [16.3.0] - 2025-08-07
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/typography` [16.1.5 ~> 16.2.0], `@semcore/icon` [16.4.0 ~> 16.5.0], `@semcore/base-components` [16.1.2 ~> 16.2.0]).
+
 ## [16.2.4] - 2025-07-10
 
 ### Fixed