+++ semcore/core/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.0.5] - 2025-08-08
+## [16.1.0] - 2025-07-24
+
+### Added
+
+- New tokens for FeaturePopover's new theme styles: `feature-popover-bg`, `feature-popover-bg-neutral`, `feature-popover-dot-neutral`, `feature-popover-dot-neutral-outer-border`.
 
 ### Fixed